+++ Centroid.py
@@ -3,7 +3,6 @@
 from tqdm import tqdm
 
 class Centroid:
-
     def __init__(self, filepath):
         self.filepath = filepath
         self.data = {}
