+++ hit_based_geometry/print_hits.py
